==> 02. Scrapped_Data_Flatten.ipynb <==
import json
import os
import io

Data_Folder = 'Data'
Final_Apps_Data = '{}\\{:02d}_Scrapped_Final_App_Data.json'.format(Data_Folder,4)

def Load_Json(file_name):
    data_dict = []
    if os.path.exists(file_name):
        with io.open(file_name, encoding='utf8') as data_file:
            data_dict = json.load(data_file)
    return data_dict

# Load JSON file into dictionnary
app_data = Load_Json(Final_Apps_Data)

# Number of scrapped apps 
len(app_data)

# Example of the data to look at at it
first_app = app_data[list(app_data.keys())[0]]
first_app

# Check available features of the apps and store their type
features = {}

for _, app in app_data.items():
    for f in app.items():
        # Do not add empty features as their data type is not defined 
        if f[1] != None:
            features[f[0]] = str(type(f[1]))

# Full list of features
for f,t in sorted(features.items()):
    print(f,t)

# Features with extra hierarchy
for f,t in features.items():
    if t == "<class 'list'>":
        print(f,t)

# 'comments' and 'screenshots' are some separate field of exploration, but 'histogram' could possibly be used in dataset 
# So: 
# 'histogram' field will be flattened
# 'comments' and 'screenshots' will go to their separate datasets
# Let's see historgam example:
hier_fields = {'comments','screenshots','histogram'}

first_app['histogram']


# We can represent it as 5 different fields
rating_example = {'rating' + str(num_stars+1):star_rating for num_stars, star_rating in enumerate(first_app['histogram'])}
rating_example

# All features except of list type with quotes around their names
for f,t in features.items():
    if t != "<class 'list'>":
        print("'"+f+"',")

# Example of the data to look at at it
first_app_reduced = {key:first_app[key] for key in set(first_app.keys()) - hier_fields if key in first_app}
first_app_reduced


# Add flatten rating histogram to the dictionnary
first_app_reduced.update(rating_example)

first_app_reduced.keys()

# Let's filter out fields that are not important for future analysis
# For that let's visualise what all fields represent
for f in sorted(set(features) - hier_fields):
    for _, app in app_data.items():
        if app[f] != None:
            print({f:app[f]})
            break

# Let's see some detailed info for some fields

# 'genre' and 'genreId' difference
comp = {}
for _, app in app_data.items():
    comp[app['genre']] = app['genreId']
comp

# 'installs' and 'minInstalls' difference
comp = {}
for _, app in app_data.items():
    comp[app['installs']] = app['minInstalls']
comp

# 'installs' and 'minInstalls' difference
comp = {}
for _, app in app_data.items():
    comp[app['currency']] = app['currency']
comp

# We can filter out the following fields:
# 'androidVersionText', - duplicates 'androidversion'
# 'developerEmail' - not needed
# 'developerInternalID' - duplicates 'developerID'
# 'descriptionHTML' - duplicates 'description'
# 'recentChangesHTML' - duplicates 'recentChanges'
# 'summaryHTML' - duplicates 'summary'
# 'installs' - duplicates 'minInstalls'
# 'genre' - duplicates 'genreId'

tech_fields = {'androidVersionText','developerEmail','developerInternalID','descriptionHTML','recentChangesHTML','summaryHTML','installs','genre'}

# Let's prepare new dataset, without extra hierarchy and without technical fields
new_dataset = []

for _, app in app_data.items():
    app_features_reduced = {key:app[key] for key in set(app.keys()) - hier_fields - tech_fields if key in app}
    if app['histogram'] != None:
        rating_feature = {'rating' + str(num_stars+1):star_rating for num_stars, star_rating in enumerate(app['histogram'])}
        app_features_reduced.update(rating_feature)
    new_dataset.append(app_features_reduced)

len(new_dataset)

# Example what we've got as a new, flatten and reduced dataset.
new_dataset[0]

# Check available features of the apps and store their type
features = {}

for app in new_dataset:
    for f in app.items():
        # Do not add empty features as their data type is not defined 
        if f[1] != None:
            features[f[0]] = str(type(f[1]))

# Next step we will split the dataset and save it to three different flat csv files
# app_base_data - dataset with most of the fields that we've defined above
# app_screenshots - dataset with user screenshot links 
# app_comments - dataset with user comments
import pandas as pd

# Convert data dictionnary to pandas dataframe and save it to csv file for further access
app_base_data = pd.DataFrame(new_dataset)

app_base_data.head()

# Write to file
app_base_data.to_csv('{}//{:02d}_app_base_data.csv'.format(Data_Folder,5), index = False)

# Let's check that csv is OK
test = pd.read_csv('{}//{:02d}_app_base_data.csv'.format(Data_Folder,5))
test.head()

test.head()

# Now screenshots dataset
screens = []
for _, app in app_data.items():
    if app['screenshots'] != None:
        for s in app['screenshots']:
            screens.append((app['appId'],s))

screens[0]

len(screens)

# Convert to pandas dataframe
app_screens = pd.DataFrame(screens)

app_screens.head()

# Write to file
app_screens.to_csv('{}\\{:02d}_app_screens.csv'.format(Data_Folder,6), index = False)

# Finally, comments dataset
comments = []
for _, app in app_data.items():
    if app['comments'] != None:
        for c in app['comments']:
            comments.append((app['appId'],c))

comments[0]

len(comments)

# Convert to pandas dataframe
app_comments = pd.DataFrame(comments)

app_comments.head()

# Write to file
app_comments.to_csv('{}\\{:02d}_app_comments.csv'.format(Data_Folder,7), index = False)
