==> 03. Base_Data_Preparation.ipynb <==
import pandas as pd
import numpy as np

# Load fetched data
Data_Folder = 'Data'
app_data = pd.read_csv('{}\\{:02d}_app_base_data.csv'.format(Data_Folder,5))

# Show the data shape (rows, columns):
app_data.shape

# Show example (top 5) of the data with all columns displayed
pd.set_option('display.max_columns', None)
app_data.head()

# Data columns info (type and number of null values)
app_data.info()

# As we can see, there are many fields with type 'object'. 
# For machine learning tasks we should transform their types to something more manageable

# For the categories identification we should see how many distinct values some feature can have
# Roughly, let's say, if number of distinct values throughout the dataset is less than 100, 
# then we could try cosider them as a categories and apply some further encoding
# Let's see the categorical variables that we could transform
for col in app_data:
    print(col)
    if len(app_data[col].unique()) < 100:
        print(app_data[col].unique())
    print()

Now we can segregate our fields by the type of required further processing:

1. As there is not enough data, we will drop the following columns: ['contentRatingDescription','originalPrice','saleText','saleTime']

2. These columns do not make much sense in our classification, so we will drop them as well:
['developer',''developerAddress','developerId','developerWebsite','privacyPolicy','url','video','videoImage']

3. We will convert to binary the following fields: 
['adSupported','containsAds','free','offersIAP','sale']

4. Field should be converted to ordinal int type in the format 'yyyymmdd' so that it is still readable:
['released','updated']

5. Filed should be formatted and converted to proper numeric:
['size']

6. Following columns need to be converted to percentage from total because of high correlation:
['rating1','rating2','rating3','rating4','rating5']

7. We need to look deeper at columns: 
['currency','version','androidVersion']

8. We will use one hot encoding for the following fields:
['contentRating','genreId']

9. We will not convert:
['appId','description','headerImage','icon','price','ratings','recentChanges','reviews','summary','title','score','minInstalls'] 


# 1. As there is not enough data, we will drop the following columns: 
# ['contentRatingDescription','originalPrice','saleText','saleTime']
app_data.drop(['contentRatingDescription','originalPrice','saleText','saleTime'], axis=1, inplace=True)

# 2. These columns does not make much sense in our classification, so we will drop them as well:
# ['developer',''developerAddress','developerId','developerWebsite','privacyPolicy','url','video','videoImage']
app_data.drop(['developer','developerAddress','developerId','developerWebsite','privacyPolicy','url','video','videoImage'], axis=1, inplace=True)

# 3. We will convert to binary the following fields: 
# ['adSupported','containsAds','free','offersIAP','sale']
app_data.replace({'adSupported':{True:1, np.nan:0}}, inplace=True)
app_data.replace({'containsAds':{True:1, np.nan:0}}, inplace=True)
app_data.replace({'offersIAP':{True:1, np.nan:0}}, inplace=True)
app_data.replace({'sale':{True:1, False:0}}, inplace=True)

# For the field 'free' it is not clear how to treat NaN values. Let's see value counts:

app_data['free'].value_counts(dropna=False)

# There are not many NaN values. We'll convert them to True as this is a major class
app_data.replace({'free':{True:1, np.nan:0,False:0}}, inplace=True)

# 4. Field should be converted to ordinal int type in the format 'yyyymmdd' so that it is still readable:
# ['released','updated']
app_data['released'] = pd.to_datetime(app_data['released'], format="%b %d, %Y").dt.strftime("%Y%m%d")
app_data['updated'] = pd.to_datetime(app_data['updated'], unit="s").dt.strftime("%Y%m%d")


# Fill in missing release date with update date
app_data['released'].fillna(app_data['updated'], inplace=True) 
app_data['updated'].fillna(app_data['released'], inplace=True) 

app_data[['released','updated']].info()

# There are still NaN values, let's see them
app_data[app_data.released.isna()]

# Let's drop these empty records
app_data.drop(app_data[app_data.released.isna()].index, inplace=True)

# Finally, convert to int
app_data['released'] = app_data['released'].astype(int)
app_data['updated'] = app_data['updated'].astype(int)

# 5. Filed should be formated and converted to proper numeric:
# ['size']
# Let's see the range of values

app_data['size'].value_counts(dropna=False)

# There are too many different values, so let's see the reduced set. 
# We'll exclude all numbers with or without dot or comma and with 'k' or 'M' at the end
app_data['size'].str.replace(r'^[\d]+[\,\.]*[\d]+[kM]+', 'matched').unique()

# 'Varies with device', 'er' and nan are not defined sizes, so let's make them all NaN at first
app_data['size'].replace({'Varies with device':np.nan,'er':np.nan}, inplace=True)

# Let's go down to kilobytes
app_data['size'] = \
    app_data['size'].str.replace(r'[kM]', '').str.replace(r'\,', '').astype(float) \
    * \
    app_data['size'].str.extract(r'([kM])', expand=False).replace([np.nan,'k','M'], [1,1,1024]).astype(int)

# Finally, replace all NaNs with zeroes
app_data['size'].replace({np.nan:0}, inplace=True)

# 6. Following columns need to be converted to percentage from total because of high correlation:
# ['rating1','rating2','rating3','rating4','rating5']
app_data['rating1'] = app_data['rating1'] / app_data['ratings']
app_data['rating2'] = app_data['rating2'] / app_data['ratings']
app_data['rating3'] = app_data['rating3'] / app_data['ratings']
app_data['rating4'] = app_data['rating4'] / app_data['ratings']
app_data['rating5'] = app_data['rating5'] / app_data['ratings']

# We cannot analyse apps that were not rated for the purpose of finding out the sources of popularity.
# So let's drop all records where 'ratings' is zero
app_data.drop(app_data[(app_data.ratings.isna())|(app_data.ratings==0)].index, inplace=True)

# 7. We need to look deeper at columns: 
# ['currency','version','androidVersion']
# First, let's look at 'currency'
app_data['currency'].value_counts(dropna=False)

# There are not many values other than 'USD' let's look if they mean something
pd.set_option('display.max_rows', None)
app_data[app_data['currency'] != 'USD']['price'].unique()

# It seems that they are referring to zero price anyway
# So, 'currency' column can be simply dropped
app_data.drop(['currency'], axis=1, inplace=True)

# Now let's look at 'version'
len(app_data['version'].unique())

# There are too many different versions. 
# It seems that without proper app history, this information will be useless for the classification
# Let's drop this field as well
app_data.drop(['version'], axis=1, inplace=True)

# Lastly, let's explore 'androidVersion' 
app_data['androidVersion'].value_counts(dropna=False)

# '4.4W' version stands for 'wearables' version of android, 
# so in order to keep apps from the same platform, let's skip those rows
app_data.drop(app_data[app_data['androidVersion'] == '4.4W'].index, inplace=True)

# NaN and 'Varies' values do not specify the exact version, so, let them be zero
app_data['androidVersion'].replace({'Varies':'0', np.nan:'0'}, inplace=True)

# Now we need to convert version to ordinal int
# To make them ordinal we'll expand version to 3 digits - adding trailing zero where needed
# And lastly, get rid of dots
app_data['androidVersion'] = app_data['androidVersion'].str.replace(r'^(\d\.\d)$', r'\1.0')
app_data['androidVersion'] = app_data['androidVersion'].str.replace(r'\.','').astype(int)
app_data['androidVersion'].value_counts(dropna=False)

# 8. We will use one hot encoding for the following fields:
# ['contentRating','genreId']
# First let's look at 'contentRating'
app_data['contentRating'].value_counts(dropna=False)

# According to Google content rating policy, 
# 'Unrated' and NaN should be treated as the highest required rating, 
# I.e. 'Adults only 18+'
app_data['contentRating'].replace({'Unrated':'Adults only 18+',np.nan:'Adults only 18+'}, inplace = True)

# Apply one hot encoding
app_data = pd.get_dummies(app_data, columns=['contentRating'])

# Now let's look at 'genreId'
app_data['genreId'].value_counts(dropna=False)

# Apply one hot encoding
app_data = pd.get_dummies(app_data, columns=['genreId'])

# Let's see how much missing data we still have
missing_data = app_data.isnull().sum().sort_values(ascending=False)
missing_data[missing_data>0]


# NaN price will fill with 0
app_data.price.fillna(0, inplace=True)

# Now let's see what we've got as a result
app_data.info()

# Let's save the dataset
app_data.to_csv('{}\\{:02d}_app_data_processed.csv'.format(Data_Folder,8), index = False)
