==> 04. Reviews_Download.ipynb <==
# Web scrape user reviews for the list for the apps we've collected previously

import os 
import time
import pandas as pd
from tqdm.auto import tqdm
from google_play_scraper import Sort, reviews_all

Data_Folder = 'Data'
apps = pd.read_csv('{}\\{:02d}_app_data_processed.csv'.format(Data_Folder,8))[['appId','reviews']]

# Reset old instances of tqdm to make it work properly
try:
    while len(tqdm._instances) > 0:
        tqdm._instances.pop().close()
except: 
    pass

unsuccessfull_reads = 0

# Run through the list of app Ids and load all user reviews
for app_Id in tqdm(list(apps[apps.reviews > 0].appId)):
    file_to = ''.join([Data_Folder,'\\{:02d}_ReviewsCSV\\'.format(9),app_Id,'.csv'])
    if os.path.exists(file_to):
        print("Skipping: " + app_Id + " " * (100 - len(app_Id)) ,end='\r')
    else:
        try:
            print("Reading: " + (app_Id + " " * (128 - len(app_Id)))[:60] + 
                  " Reviews: " + str(apps[apps.appId == app_Id].reviews.values.item()) + "    "
                  ,end='\r')

            df = pd.DataFrame(reviews_all(app_Id))
            df.to_csv(file_to, index = False)
        except:
            unsuccessfull_reads += 1
    
print("Done. Unsuccessfull reads: {}".format(str(unsuccessfull_reads)))
