==> 05. Screenshots_Download.ipynb <==
import requests
import re
import os 
from pathlib import Path
import pandas as pd
from tqdm import tqdm


Data_Folder = 'Data'

app_screens = pd.read_csv('{}\\{:02d}_app_screens.csv'.format(Data_Folder,6))

# Reset old instances of tqdm to make it work properly

try:
    while len(tqdm._instances) > 0:
        tqdm._instances.pop().close()
except:
    pass

# Run through the list of screenshot file paths and load them into 'Screenshots' folder
for index,row in tqdm(app_screens.iterrows()):
    fileName = os.path.join(Data_Folder,'{:02d}_Screenshots'.format(10), re.split('/',row[1])[-1:][0] + '.png')
    f = Path(fileName)
    if not f.is_file():
        with open(fileName,'wb') as f:
            f.write(requests.get(row[1]).content)
