==> 06. Header_and_Icon_Download.ipynb <==
import requests
import os 
from pathlib import Path
import pandas as pd
from tqdm.auto import tqdm

Data_Folder = 'Data'
df_appPic = pd.read_csv('{}\\{:02d}_app_data_processed.csv'.format(Data_Folder,8))[['appId','headerImage','icon']]

imgSource = [('headerImage','{}\\{:02d}_HeaderImages'.format(Data_Folder, 11)),
             ('icon','{}\\{:02d}_Icons'.format(Data_Folder,12))]

# Run through the list of appIds and load corresponding header and icon image files into their respective folders
for appId in tqdm(list(df_appPic['appId'])):
    for src in imgSource:
        
        try:
            fileName = os.path.join(src[1], appId + '.png')
            f = Path(fileName)
            if not f.is_file():
                with open(fileName,'wb') as f:
                    f.write(requests.get(df_appPic[df_appPic.appId == appId][src[0]].values[0]).content)
        except:
            continue


df_appPic[df_appPic.appId == appId][src[0]].values[0]
