==> 08. Reviews_Convert_to_Pickle.ipynb <==
import pandas as pd
import os
from tqdm.auto import tqdm

Data_Folder = 'Data'
df_appId = pd.read_csv('{}\\{:02d}_app_data_processed.csv'.format(Data_Folder,8), usecols=['appId'])

# Data folder with user reviews has size 126Gb 
# It will be very difficult to operate with this data in future. 
# Let's store all reviews in another folder in python-native format 'pickle'
# As a bonus, pickle will store datatypes for the columns
for app_Id in tqdm(list(df_appId['appId'])):
    try:        
        file_from = ''.join([Data_Folder,'\\{:02d}_ReviewsCSV\\'.format(9),app_Id,'.csv'])
        file_to = ''.join([Data_Folder,'\\{:02d}_ReviewsPKL\\'.format(13),app_Id,'.pkl'])
        if os.path.exists(file_from):
            if not os.path.exists(file_to):
                reviews = pd.read_csv(file_from, 
                                      usecols=['at','content','repliedAt','score','thumbsUpCount'],
                                      dtype = {'score':'int8','thumbsUpCount':'int8'},
                                      parse_dates=['at','repliedAt']
                                     )

                reviews.to_pickle(file_to)
    except:
        continue
