==> 09.5_Reviews_to_File_(Stats).ipynb <==
# Making a single pickle file with all statistical data 
# from 'ReviewsSentiment' folder

import pandas as pd
import os
from tqdm.auto import tqdm
from datetime import datetime

Data_Folder = 'Data'
df_appId = pd.read_csv('{}\\{:02d}_app_data_processed.csv'.format(Data_Folder,8), usecols=['appId'])

file_to_csv = '{}\\{:02d}_app_all_review_stats.csv'.format(Data_Folder,20)
file_to_pkl = '{}\\{:02d}_app_all_review_stats.pkl'.format(Data_Folder,21)

# For easier statistical plotting we need to consolidate the data and simplify in order to optimize memory allocation
# Date fields are shorten, appId is replaced with dataframe index 
# Running through the list of appIds and loading app review stats fields 
# Saving into a separate csv file as it allowes append writing, contrary to pickle
try:
    while len(tqdm._instances) > 0:
        tqdm._instances.pop().close()
except:
    pass

try:
    processed_ids = pd.read_csv(file_to_csv, usecols=['id'])
    last_processed_id = df_appId.loc[processed_ids['id'].max()]['appId']
    mode = 'a'
except:
    last_processed_id = ''
    mode = 'w'

for app_Id in tqdm(list(df_appId[df_appId['appId']>last_processed_id]['appId'])):
        
    file_from = ''.join([Data_Folder,'\\{:02d}_ReviewsSentiment\\'.format(19),app_Id,'.pkl'])
    
    if not os.path.exists(file_from):
        continue

    scores = pd.read_pickle(file_from)

    scores['id'] = df_appId.index[df_appId.appId == app_Id][0]
    scores['origId'] = scores.index
    scores['at'] = pd.to_datetime(scores['at'], format="%Y-%m-%d").dt.strftime("%Y%m%d")
    scores['repliedAt'] = scores.apply(lambda x: 0 if pd.isnull(x['repliedAt']) else pd.to_datetime(x['repliedAt'], format="%Y-%m-%d").strftime("%Y%m%d"), axis=1)
    scores['reviewLength'] = scores['content'].str.len()

    scores.to_csv(file_to_csv, 
                columns=['id','origId','at','repliedAt','score','thumbsUpCount','polarity','subjectivity','reviewLength'], 
                index = False,
                header = (mode == 'w'),
                mode = mode)
    mode = 'a'

# Reading 'score' dataframe
df_stats = pd.read_csv(file_to_csv)

# Converting date fields back to date
df_stats['at'] = pd.to_datetime(df_stats['at'],format='%Y%m%d', errors='ignore')
df_stats['repliedAt'] = pd.to_datetime(df_stats['repliedAt'],format='%Y%m%d', errors='coerce')

# Save it altogether as pickle for keeping the data types
df_stats.to_pickle(file_to_pkl)

# Read stats dataframe to check its structure
df_stats = pd.read_pickle(file_to_pkl)

df_stats.info()
