==> 09.6_Reviews_Bag_of_Words.ipynb <==
import pandas as pd
import os
from tqdm.auto import tqdm
import collections

Data_Folder = 'Data'
df_appId = pd.read_csv('{}\\{:02d}_app_data_processed.csv'.format(Data_Folder,8), usecols=['appId'])

# Let's collect our lemmas into bag of words
try:
    while len(tqdm._instances) > 0:
        tqdm._instances.pop().close()
except:
    pass

dictionary = {}

def text2bow(words):
    word_frequences = collections.defaultdict(int)
    for word in words:
        if word not in dictionary:
            dictionary[word] = len(dictionary)
        word_frequences[dictionary[word]] += 1
 
    return list(word_frequences.items())

for app_Id in tqdm(list(df_appId['appId'])):
        
    file_from = '{}\\{:02d}_ReviewsLemma\\{}.pkl'.format(Data_Folder,18,app_Id)
    file_to = '{}\\{:02d}_ReviewsBoW\\{}.pkl'.format(Data_Folder,22,app_Id)
    if not os.path.exists(file_to):
        if os.path.exists(file_from):
            scores = pd.read_pickle(file_from)
            scores['content'] = scores['content'].apply(text2bow)
            scores.to_pickle(file_to)

bow_dict = pd.DataFrame(dictionary.items())
bow_dict.columns = ['word','id']
bow_dict.to_pickle('{}\\{:02d}_BoW_Dictionary.pkl'.format(Data_Folder,23))
